fixed merge issue with previous issue
